+++ src/admin/plugin-page.php
@@ -10,6 +10,7 @@ declare(strict_types=1);
 namespace TotalMailQueue\Admin;
 
 use TotalMailQueue\Admin\Pages\SmtpPage;
+use TotalMailQueue\Admin\Pages\SourcesPage;
 use TotalMailQueue\Admin\Tables\LogTable;
 use TotalMailQueue\Database\Schema;
 use TotalMailQueue\Plugin;
@@ -33,6 +34,7 @@ final class PluginPage {
 	public const TAB_LOG       = 'wp_tmq_mail_queue-tab-log';
 	public const TAB_QUEUE     = 'wp_tmq_mail_queue-tab-queue';
 	public const TAB_SMTP      = 'wp_tmq_mail_queue-tab-smtp';
+	public const TAB_SOURCES   = 'wp_tmq_mail_queue-tab-sources';
 	public const TAB_CRON_INFO = 'wp_tmq_mail_queue-tab-croninfo';
 	public const TAB_FAQ       = 'wp_tmq_mail_queue-tab-faq';
 
@@ -62,6 +64,9 @@ final class PluginPage {
 				break;
 			case self::TAB_SMTP:
 				SmtpPage::render();
+				break;
+			case self::TAB_SOURCES:
+				SourcesPage::render();
 				break;
 			case self::TAB_FAQ:
 				self::renderFaqTab();
@@ -134,6 +139,7 @@ final class PluginPage {
 			self::TAB_LOG      => __( 'Log', 'total-mail-queue' ),
 			self::TAB_QUEUE    => __( 'Retention', 'total-mail-queue' ),
 			self::TAB_SMTP     => __( 'SMTP Accounts', 'total-mail-queue' ),
+			self::TAB_SOURCES  => __( 'Sources', 'total-mail-queue' ),
 		);
 		if ( defined( 'DISABLE_WP_CRON' ) && DISABLE_WP_CRON ) {
 			$tabs[ self::TAB_CRON_INFO ] = __( 'Cron Information', 'total-mail-queue' );